+++ src/TextShow/Data/Ord.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TemplateHaskell #-}
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-|
@@ -14,12 +15,27 @@ Portability: GHC
 -}
 module TextShow.Data.Ord () where
 
-import GHC.Exts (Down)
-import TextShow.TH.Internal (deriveTextShow, deriveTextShow1)
+import GHC.Exts (Down(..))
+
+import TextShow.Classes ( TextShow(..), TextShow1(..)
+                        , showbPrec1, showbUnaryWith )
+import TextShow.TH.Internal (deriveTextShow)
+
+-- | This instance would be equivalent to a derived 'TextShow' instance
+-- if the 'getDown' field were removed.
+--
+-- /Since: 2/
+instance TextShow a => TextShow (Down a) where
+    showbPrec = showbPrec1
+    {-# INLINE showbPrec #-}
+
+-- | This instance would be equivalent to a derived 'TextShow1' instance
+-- if the 'getDown' field were removed.
+--
+-- /Since: 2/
+instance TextShow1 Down where
+    liftShowbPrec sp _ p (Down x) = showbUnaryWith sp "Down" p x
+    {-# INLINE liftShowbPrec #-}
 
 -- | /Since: 2/
-$(deriveTextShow  ''Ordering)
--- | /Since: 2/
-$(deriveTextShow  ''Down)
--- | /Since: 2/
-$(deriveTextShow1 ''Down)
+$(deriveTextShow ''Ordering)