+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### next [????.??.??]
+* Update the `TextShow(1)` instances for `Down` and `ThreadId` to match their
+  implementations in `base-4.14`.
+
 ### 3.8.3 [2019.11.06]
 * Use `base-compat-batteries-0.11.0` to define instances for `(:~:)` back to
   `base-4.5` and instances for `(:~~:)` back to `base-4.9`.